NFT serial type changed

+++ contracts/common/BaseHTS.sol
@@ -32,15 +32,9 @@ contract BaseHTS is HederaTokenService, IBaseHTS {
         address token,
         address sender,
         address receiver,
-        int256 serial
+        int64 serial
     ) external override returns (int256) {
-        return
-            HederaTokenService.transferNFT(
-                token,
-                sender,
-                receiver,
-                int64(serial)
-            );
+        return HederaTokenService.transferNFT(token, sender, receiver, serial);
     }
 
     function associateTokenPublic(
